Changing lua calls
from "Update" "Render" to "OnUpdate" "OnRender"
the reason for this was when you had a gameObject in lua, and needed to
call its "notifyColission" function. It will have issues with the ones
registered in lua and the c++ one. For now. obj:Call   means call the
c++ functions and obj:OnCall means the lua call.