Possible deletion while running avoid.
This actually caused a crash in Dogurai. Now its fixed.

This happens when during a iteration, some object is added to the pool and forcing it to reallocate.

Now at this point, the old pointer to the list is added to a vector and at the begin of the next game loop those references are erased.